validate fields and check ownership before editing and deleting

+++ app/Http/Controllers/ClimbController.php
@@ -3,11 +3,19 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Validation\Rule;
 use App\Climb;
-use Gate;
+use App\User;
+use Auth;
+use JWTAuth;
+use Tymon\JWTAuth\Exceptions\JWTException;
 
 class ClimbController extends Controller
-{
+{   
+    public function __construct()
+    {
+        $this->middleware('auth:api', ['only' => ['store', 'update', 'destroy']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -49,13 +57,30 @@ class ClimbController extends Controller
      */
     public function store(Request $request)
     {
+        $this->validate($request, [
+            'name' => 'required|max:255',
+            'description' => 'required|max:255',
+            'length' => 'nullable|integer',
+            'rating' => 'required|string',
+            'type' => ['required', Rule::in(['Top Rope', 'Trad', 'Sport', 'Boulder'])],
+            'gear_needed' => 'nullable|string',
+            'added_by' => 'string',
+            'location' => 'string',
+            'public' => Rule::in(['1', '0'])
+        ]);
+
+        
+        if (!$user = JWTAuth::parseToken()->authenticate()) {
+            return response()->json(['msg' => 'User not found'], 404);
+        }
+
         $name = $request->input( 'name');
         $description = $request->input('description');
         $length = $request->input('length');
         $rating = $request->input('rating');
         $type = $request->input('type');
         $gear_needed = $request->input( 'gear_needed');
-        $added_by = $request->input( 'added_by');
+        $added_by = $user->email;
         $location = $request->input( 'location');
         $public = $request->input( 'public');
         
@@ -78,12 +103,12 @@ class ClimbController extends Controller
                 'href' => '/api/v1/climbs/' . $climb->id,
                 'method' => 'GET'
             ];
-            $response = ['msg' => 'climb created', 'climb' => $climb];
+            $response = ['msg' => 'The climb "' . $climb->name . '" was created', 'climb' => $climb];
             return response()->json($response, 201);
         }
 
         $response = ['msg' => 'An error occured.'];
-        return response()->json($response, 404);
+        return response()->json($response, 400);
 
     }
 
@@ -106,17 +131,7 @@ class ClimbController extends Controller
         return response()->json($response, 200);        
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function edit($id)
-    {
-        //
-    }
-
+   
     /**
      * Update the specified resource in storage.
      *
@@ -126,37 +141,61 @@ class ClimbController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $this->validate($request, [
+            'name' => 'max:255',
+            'description' => 'max:255',
+            'length' => 'nullable|integer',
+            'rating' => 'string',
+            'type' => Rule::in(['Top Rope', 'Trad', 'Sport', 'Boulder']),
+            'gear_needed' => 'nullable|string',
+            'added_by' => 'string',
+            'location' => 'string',
+            'public' => Rule::in(['1', '0'])
+        ]);
+        try {
+            
+            $user = JWTAuth::parseToken()->authenticate();
+            
+        } catch (JWTException $e) {
+
+            return response()->json(['error' => 'User not found'], 401);
+
+        }
+                
+        
         $climb = Climb::where('id', $id)->first();
 
-        // if (Gate::denies('manipulate-climb', $climb)) {
-        //     $response = ['msg' => 'User is not authorized to edit climb.'];
-        //     return response()->json($response, 404);
-        // }
-        $climb->name = $request->input('name') ?? $climb->name;
-        $climb->length = $request->input('length') ?? $climb->length;
-        $climb->description = $request->input('description') ?? $climb->description;
-        $climb->rating = $request->input('rating') ?? $climb->rating;
-        $climb->gear_needed = $request->input('gear_needed') ?? $climb->gear_needed;
-        $climb->type = $request->input('type') ?? $climb->type;
-        $climb->location = $request->input('location') ?? $climb->location;
-        $climb->added_by = $request->input('added_by') ?? $climb->added_by;
 
+        if ($user->email == $climb->added_by || $climb->public) {
 
-        
+            $climb->name = $request->input('name') ?? $climb->name;
+            $climb->length = $request->input('length') ?? $climb->length;
+            $climb->description = $request->input('description') ?? $climb->description;
+            $climb->rating = $request->input('rating') ?? $climb->rating;
+            $climb->gear_needed = $request->input('gear_needed') ?? $climb->gear_needed;
+            $climb->type = $request->input('type') ?? $climb->type;
+            $climb->location = $request->input('location') ?? $climb->location;
+            $climb->added_by = $user->email ?? $climb->added_by;
+    
+            if ($climb->save()) {
+    
+                $climb->view_climb = [
+                    'href' => '/api/v1/climbs/' . $climb->id,
+                    'method' => 'GET'
+                ];
 
-        if ($climb->save()) {
-
-            $climb->view_climb = [
-                'href' => '/api/v1/climbs/' . $climb->id,
-                'method' => 'GET'
-            ];
-            $response = ['msg' => $climb->name . ' updated', 'climb' => $climb];
-            return response()->json($response, 201);
+                $response = ['msg' => 'The climb "' . $climb->name . '" was updated', 'climb' => $climb];
+                return response()->json($response, 201);
+            
+            } else {
+             
+                $response = ['msg' => 'An error occured.'];
+                return response()->json($response, 400);
+            } 
+        } else {
+            $response = ['msg' => 'User does not have sufficient permissions to edit "' . $climb->name . '"'];
+            return response()->json($response, 403);        
         }
-
-
-        $response = ['msg' => 'An error occured.'];
-        return response()->json($response, 404);
     }
 
     /**
@@ -167,18 +206,35 @@ class ClimbController extends Controller
      */
     public function destroy($id)
     {
-        $climb = Climb::where('id', $id)->firstOrFail();
-        if ($climb->delete()) {
-            $response = [
-                'msg' => 'Climb Deleted',
-                'create' => [
-                    'href' => '/api/v1/climb',
-                    'method' => 'POST',
-                    'params' => 'name, description, length, location, added_by, gear_needed, rating, type'
-                ]
-            ];
+        try {
 
-            return response()->json($response, 200);      
+            $user = JWTAuth::parseToken()->authenticate();
+        } catch (JWTException $e) {
+
+            return response()->json(['msg' => 'User not found'], 401);
         }
+
+        $climb = Climb::where('id', $id)->firstOrFail();
+        if ($user->email == $climb->added_by && !$climb->public) {
+            if ($climb->delete()) {
+                $msg = 'The climb "' . $climb->name . '" was deleted.'; 
+                $response = [
+                    'msg' => $msg,
+                    'create' => [
+                        'href' => '/api/v1/climb',
+                        'method' => 'POST',
+                        'params' => 'name, description, length, location, added_by, gear_needed, rating, type'
+                    ]
+                ];
+
+                return response()->json($response, 200);
+            }      
+        } else {
+            $response = ['msg' => 'User does not have sufficient permissions to delete "' . $climb->name . '"'];
+            return response()->json($response, 403);
+        }
+        $response = ['msg' => 'Something went wrong.'];
+        return response()->json($response, 400);
+        
     }
 }
